Importer > Mesh: support looping and clips for parent sequences of the paint layer, recursively.
This computes the local time in the base animation sequence of the paint layer taking into account all the parents' sequences looping and clips, and the groups visibility spans.
It is still not the full model because transform key frames that are inside a looping sequence or clip are only imported as the base animation.