Add properties to objects created during import, and export the original Quill objects during export.
When importing a Quill scene the created Blender objects now have properties to remember the scene, layer and drawing index they were created from.
When exporting, if we find that the mesh was created from a Quill import, we re-open the original scene and use the original Quill drawing instead of exporting a wireframe.